+++ image/image-local-features/src/main/java/org/openimaj/feature/local/matcher/FastEuclideanKeypointMatcher.java
@@ -77,11 +77,11 @@ public class FastEuclideanKeypointMatcher<T extends Keypoint> implements LocalFe
 			data[i] = keys1.get(i).ivec;
 		
 		int [] argmins = new int[keys1.size()];
-		int [] mins = new int[keys1.size()];
+		float [] mins = new float[keys1.size()];
 		modelKeypointsKNN.searchNN(data, argmins, mins);
 		
 		for (int i=0; i<keys1.size(); i++) {
-			int distsq = mins[i];
+			float distsq = mins[i];
 			
 			if (distsq < threshold) {
 				matches.add(new Pair<T>(keys1.get(i), modelKeypoints.get(argmins[i])));
