+++ image/image-local-features/src/main/java/org/openimaj/feature/local/matcher/FastLimitedEuclideanKeypointMatcher.java
@@ -74,9 +74,9 @@ public class FastLimitedEuclideanKeypointMatcher<T extends Keypoint> implements 
 	}
 
 	class WPair extends Pair<T> implements Comparable<WPair> {
-		int weight;
+		float weight;
 
-		public WPair(T obj1, T obj2, int weight) {
+		public WPair(T obj1, T obj2, float weight) {
 			super(obj1, obj2);
 			this.weight = weight;
 		}
@@ -98,11 +98,11 @@ public class FastLimitedEuclideanKeypointMatcher<T extends Keypoint> implements 
 			data[i] = keys1.get(i).ivec;
 		
 		int [] argmins = new int[keys1.size()];
-		int [] mins = new int[keys1.size()];
+		float [] mins = new float[keys1.size()];
 		modelKeypointsKNN.searchNN(data, argmins, mins);
 		
 		for (int i=0; i<keys1.size(); i++) {
-			int distsq = mins[i];
+			float distsq = mins[i];
 			
 			mq.add(new WPair(keys1.get(i), modelKeypoints.get(argmins[i]), distsq));
 		}
